更改校验与增加黑名单校验

+++ qinjee-parent/qinjee-masterdata/src/main/java/com/qinjee/masterdata/model/entity/UserArchive.java
@@ -177,7 +177,7 @@ public class UserArchive implements Serializable {
      * 工龄
      */
     @ApiModelProperty("工龄")
-    private BigDecimal workingPeriod;
+    private Integer workingPeriod;
 
     /**
      * 试用到期时间
